Remove unecessary css class, move component css to its own css module and refactoring. Bugzilla 1306494
