+++ src/components/Diff/styles.module.css
@@ -1,0 +1,13 @@
+.addition {
+  color: seagreen;
+  background-color: palegreen;
+}
+
+.deletion {
+  color: red;
+  background-color: mistyrose;
+}
+
+.unchanged {
+  color: grey;
+}
